% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_comb.R
\name{csbase}
\alias{csbase}
\title{Base forecast component}
\usage{
csbase(
  base,
  agg_mat = NULL,
  block_diag = "fr",
  p = NULL,
  comb = "ols",
  res = NULL,
  ...
)
}
\arguments{
\item{base}{A list of \eqn{p} (\eqn{h \times n}) numeric matrix or multivariate time series
(\code{mts} class) containing the base forecasts to be reconciled; \eqn{h} is the forecast
horizon, and \eqn{n} is the total number of time series (\eqn{n = n_a + n_b}).}

\item{agg_mat}{A (\eqn{n_a \times n_b}) numeric matrix representing the cross-sectional
aggregation matrix. It maps the \eqn{n_b} bottom-level (free)
variables into the \eqn{n_a} upper (constrained) variables.}

\item{block_diag}{A string specifying the block-diagonal structure of the covariance matrix:
\code{fr} reconciliation structure, \code{fc} combination structure, \code{none} full covariance matrix.}

\item{comb}{A string specifying the reconciliation method. For a complete list, see [cscov].}

\item{res}{A list of \eqn{p} (\eqn{N \times n}) optional numeric matrix containing the
in-sample residuals. This matrix is used to compute some covariance matrices.}

\item{...}{Arguments passed on to \code{cscov} (\pkg{FoReco}).}
}
\value{
A (\eqn{h \times n}) numeric matrix of cross-sectional combined forecasts.
}
\description{
Base forecast component
}
