% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projmat.R
\name{csprojmat}
\alias{csprojmat}
\title{Projection matrix}
\usage{
csprojmat(
  agg_mat,
  cons_mat,
  block_diag = "fr",
  p = NULL,
  comb = "ols",
  res = NULL,
  approach = "proj",
  ...
)
}
\arguments{
\item{agg_mat}{A (\eqn{n_a \times n_b}) numeric matrix representing the cross-sectional
aggregation matrix. It maps the \eqn{n_b} bottom-level (free)
variables into the \eqn{n_a} upper (constrained) variables.}

\item{cons_mat}{A (\eqn{n_a \times n}) numeric matrix representing the cross-sectional
zero constraints. It spans the null space for the reconciled forecasts.}

\item{block_diag}{A string specifying the block-diagonal structure of the covariance matrix:
\code{fr} reconciliation structure, \code{fc} combination structure, \code{none} full covariance matrix.}

\item{comb}{A string specifying the reconciliation method. For a complete list, see [cscov].}

\item{res}{A list of \eqn{p} (\eqn{N \times n}) optional numeric matrix containing the
in-sample residuals. This matrix is used to compute some covariance matrices.}

\item{approach}{A string specifying the approach used to compute the reconciled
forecasts. Options include:
  \itemize{
  \item "\code{proj}" (\emph{default}): Projection approach according to Byron (1978, 1979).
  \item "\code{strc}": Structural approach as proposed by Hyndman et al. (2011).
  }}

\item{...}{Arguments passed on to \code{cscov} (\pkg{FoReco}).}
}
\value{
A list
}
\description{
Projection matrix
}
