% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projmat.R
\name{occmat}
\alias{occmat}
\title{Matrices for the optimal coherent forecast combination}
\usage{
occmat(agg_mat, cons_mat, p = NULL, matNA = NULL,
       comb = "ols", res = NULL, approach = "proj", ...)
}
\arguments{
\item{agg_mat}{A (\eqn{n_u \times n_b}) numeric matrix representing the cross-sectional
aggregation matrix. It maps the \eqn{n_b} bottom-level (free)
variables into the \eqn{n_u} upper (constrained) variables.}

\item{cons_mat}{A (\eqn{n_u \times n}) numeric matrix representing the cross-sectional
zero constraints. It spans the null space for the reconciled forecasts.}

\item{p}{Total number of experts.}

\item{matNA}{A (\eqn{n \times p}) matrix consisting of 0s and 1s,
where each element indicates whether expert \eqn{j} (column) has
provided a forecast for variable \eqn{i} (row). If expert \eqn{j}
has provided a forecast for variable \eqn{i}, the corresponding
element \eqn{(i,j)} is 1; otherwise, it is 0.}

\item{comb}{A string specifying the reconciliation method.
\itemize{
\item \href{https://danigiro.github.io/FoReco/}{\code{FoReco}} approximations:
\code{"ols"}, \code{"wls"}, \code{"shr"}, \code{"sam"}.
\item "\code{shrbe}"/"\code{sambe}" - shrunk/sample block diagonal covariance by experts.
\item "\code{shrbv}"/"\code{sambv}" - shrunk/sample block diagonal covariance by variables.
}}

\item{res}{A list of \eqn{p} numeric (\eqn{N \times n}) matrix containing the
in-sample residuals. This input is used to compute some covariance matrices.}

\item{approach}{A string specifying the approach used to compute the reconciled
forecasts. Options include:
\itemize{
\item "\code{proj}" (\emph{default}): zero-constrained approach.
\item "\code{strc}": Structural approach.
}}

\item{...}{Arguments passed on to
\href{https://danigiro.github.io/FoReco/reference/cscov.html}{\code{FoReco::cscov}}.}
}
\value{
A list of matrices:
\item{M}{The projection matrix.}
\item{Omega}{The matrix formed by the combination weights of the multi-task forecast combination.}
\item{W}{The forecast error covariance matrix of the base forecasts.}
\item{Wc}{The forecast error covariance matrix of the combined forecasts.}
\item{Wtilde}{The forecast error covariance matrix of the reconciled combined forecasts.}
\item{K}{The matrix that replicates a vector.}
}
\description{
This function computes the matrices required for the optimal coherent
forecast combination \link{csocc}, as described in Girolimetto and Di Fonzo (2024).
These matrices serve as the foundation for building forecasts that effectively
combines the individual information from multiple experts while ensuring
coherence across the variables.
}
\references{
Girolimetto, D. and Di Fonzo, T. (2024), Coherent forecast combination for linearly
constrained multiple time series, \emph{mimeo}.
}
\seealso{
Optimal coherent combination: 
\code{\link{cscov}()},
\code{\link{csocc}()},
\code{\link{occbase}()}
}
\concept{Optimal coherent combination}
