% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-sequential.R
\name{csscr}
\alias{csscr}
\title{Cross-sectional sequential combination-reconciliation}
\usage{
csscr(base, res = NULL, fc = "sa", mse = TRUE, shrink = TRUE, ...)
}
\arguments{
\item{base}{A list of \eqn{p} (\eqn{h \times n}) numeric matrix or multivariate time series
(\code{mts} class) containing the base forecasts to be reconciled; \eqn{h} is the forecast
horizon, and \eqn{n} is the total number of time series (\eqn{n = n_a + n_b}).}

\item{res}{A list of \eqn{p} (\eqn{N \times n}) optional numeric matrix containing the
in-sample residuals. This matrix is used to compute some covariance matrices.}

\item{fc}{A string specifying the combination method}

\item{mse}{todo}

\item{shrink}{todo}

\item{...}{Arguments passed on to \code{csrec} (\pkg{FoReco}).}
}
\value{
A (\eqn{h \times n}) numeric matrix of cross-sectional reconciled forecasts.
}
\description{
Cross-sectional sequential combination-reconciliation
}
