% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoCo2-package.R
\docType{package}
\name{FoCo2-package}
\alias{FoCo2}
\alias{FoCo2-package}
\title{FoCo2: Coherent Forecast Combination for Linearly Constrained Multiple Time Series}
\description{
Methods and tools designed to improve the forecast accuracy for a linearly constrained multiple time series, while fulfilling the linear/aggregation relationships linking the components (Girolimetto and Di Fonzo, 2024 \doi{10.48550/arXiv.2412.03429}). 'FoCo2' offers multi-task forecast combination and reconciliation approaches leveraging input from multiple forecasting models or experts and ensuring that the resulting forecasts satisfy specified linear constraints. In addition, linear inequality constraints (e.g., non-negativity of the forecasts) can be imposed, if needed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danigiro/FoCo2/}
  \item \url{https://danigiro.github.io/FoCo2/}
  \item Report bugs at \url{https://github.com/danigiro/FoCo2/issues/}
}

}
\author{
\strong{Maintainer}: Daniele Girolimetto \email{daniele.girolimetto@unipd.it} (\href{https://orcid.org/0000-0001-9387-1232}{ORCID}) [funder, copyright holder]

Authors:
\itemize{
  \item Tommaso Di Fonzo (\href{https://orcid.org/0000-0003-3388-7827}{ORCID}) [funder]
}

}
\keyword{package}
